Windows build fixes for cbr libraries. Note that bugs 91, 92, and 93 were opened because this only gets the system building.  Most fixes were simple type changes (e.g. don't use xxx_t types) and cross platform fixes such as using the correct sleep function.

+++ bench/src/TimerJitterBenchmark.cpp
@@ -82,7 +82,7 @@ void TimerJitterBenchmark::start() {
 
     SILOG(benchmark,info,
           ITERATIONS << " timer invokations, " << dur << ": "
-          << "stddev " << sqrt(diff_var) << " ns"
+          << "stddev " << sqrt((double)diff_var) << " ns"
           );
 
     notifyFinished();
